# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * KobeAI API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface LoginRequest {
  student_id: string;
  pin: string;
  device_id: string;
}

export interface LoginResponse {
  access_token: string;
  token_type: string;
  student_name: string;
  grade: string;
  wallet_balance: number;
}

export interface TeacherLoginRequest {
  email: string;
  password: string;
}

export interface TeacherLoginResponse {
  access_token: string;
  token_type: string;
  teacher_name: string;
  school: string;
  role: string;
}

export interface ParentLoginRequest {
  phone: string;
  pin: string;
}

export interface ParentLoginResponse {
  access_token: string;
  token_type: string;
  parent_name: string;
}

export interface ActivityItem {
  id: string;
  student_name: string;
  action: string;
  points: number;
  timestamp: string;
}

export interface TeacherDashboardStats {
  total_students: number;
  active_today: number;
  total_points: number;
  avg_performance: number;
  questions_today: number;
  online_devices: number;
  recent_activity: ActivityItem[];
}

export interface Student {
  id: string;
  student_id: string;
  name: string;
  grade: string;
  points: number;
  status: string;
  last_active?: string;
}

export interface StudentsListResponse {
  students: Student[];
  total: number;
}

export interface AttendanceRecord {
  id: string;
  student_id: string;
  student_name: string;
  grade: string;
  check_in_time?: string;
  status: string;
  points_earned: number;
}

export interface AttendanceListResponse {
  records: AttendanceRecord[];
  date: string;
  total_present: number;
  total_absent: number;
  total_students: number;
}

export interface LeaderboardEntry {
  rank: number;
  student_id: string;
  name: string;
  grade: string;
  points: number;
  badge?: string;
}

export interface LeaderboardResponse {
  entries: LeaderboardEntry[];
  period: string;
}

export interface StudentBalance {
  id: string;
  student_id: string;
  name: string;
  grade: string;
  balance: number;
  total_deposited: number;
  total_spent: number;
  ai_questions_spend: number;
  quiz_spend: number;
  questions_count: number;
  quizzes_count: number;
  status: string;
}

export interface BalanceSummary {
  total_accounts: number;
  total_balance: number;
  low_balance_count: number;
}

export interface StudentBalancesResponse {
  students: StudentBalance[];
  summary: BalanceSummary;
}

export interface DepositRequest {
  student_id: string;
  amount: number;
  deposit_method?: string;
  notes?: string;
}

export interface DepositResponse {
  success: boolean;
  deposit_id: string;
  receipt_number: string;
  new_balance: number;
  message: string;
}

export interface BillingSummary {
  period: string;
  total_ai_questions: number;
  total_quizzes: number;
  ai_cost: number;
  quiz_cost: number;
  subscription_fee: number;
  total_amount: number;
  status: string;
}

export interface QuizSummary {
  id: string;
  title: string;
  subject: string;
  questions_count: number;
  points_possible: number;
  duration_minutes: number;
}

export interface QuizzesListResponse {
  quizzes: QuizSummary[];
}

export interface QuizQuestion {
  id: string;
  text: string;
  options: string[];
  points: number;
}

export interface QuizStartResponse {
  attempt_id: string;
  quiz_id: string;
  title: string;
  subject: string;
  questions: QuizQuestion[];
  time_limit_minutes: number;
  total_points: number;
}

export interface QuizAnswer {
  question_id: string;
  answer: string;
}

export interface QuizSubmitRequest {
  attempt_id: string;
  answers: QuizAnswer[];
}

export interface QuizSubmitResponse {
  score: number;
  total_points: number;
  points_earned: number;
  correct_answers: number;
  total_questions: number;
  message: string;
}

export interface Transaction {
  id: string;
  amount: number;
  type: string;
  description: string;
  created_at: string;
}

export interface WalletResponse {
  balance: number;
  total_earned: number;
  level: number;
  daily_earned: number;
  daily_limit: number;
  recent_transactions: Transaction[];
}

export interface ChildInfo {
  id: string;
  name: string;
  grade: string;
  balance: number;
  today_points: number;
  total_points: number;
  attendance_streak: number;
}

export interface ParentDashboardResponse {
  parent_name: string;
  children: ChildInfo[];
}

export interface ChildActivityItem {
  id: string;
  type: string;
  description: string;
  points: number;
  timestamp: string;
  subject?: string;
}

export interface ChildActivityResponse {
  child_name: string;
  activities: ChildActivityItem[];
}

export interface ChildWalletInfo {
  id: string;
  name: string;
  grade: string;
  balance: number;
  daily_limit: number;
  transactions: Transaction[];
}

export interface ParentWalletResponse {
  total_balance: number;
  children: ChildWalletInfo[];
}

export interface AddFundsRequest {
  child_id: string;
  amount: number;
  method?: string;
}

export interface AddFundsResponse {
  success: boolean;
  new_balance: number;
  message: string;
  receipt_id: string;
}

export type GetStudentsParams = {
  grade?: string;
  search?: string;
};

export type GetAttendanceParams = {
  date?: string;
};