# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: KobeAI API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /v1/bursar/deposit:
    post:
      operationId: addDeposit
      tags: [bursar]
      summary: Add deposit to student wallet
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DepositRequest"
      responses:
        "200":
          description: Deposit successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DepositResponse"
...
components:
  schemas:
...
    DepositRequest:
      type: object
      properties:
        student_id:
          type: string
        amount:
          type: number
        deposit_method:
          type: string
        notes:
          type: string
      required:
        - student_id
        - amount
    DepositResponse:
      type: object
      properties:
        success:
          type: boolean
        deposit_id:
          type: string
        receipt_number:
          type: string
        new_balance:
          type: number
        message:
          type: string
      required:
        - success
        - deposit_id
        - receipt_number
        - new_balance
        - message
...